discordmusic/requirements.txt
discord.py==2.3.2
yt-dlp
PyNaCl==1.5.0
python-dotenv
google-api-python-client
requests